requirements.txt
easydict
omegaconf
tensorboardX
trimesh
PyMCubes
rich
torch_ema
visdom
tqdm
requests
h5py
open-clip-torch
imageio
einops
chardet
matplotlib
scikit-image
plotly
taming-transformers
kornia
clip
transformers
diffusers
datasets
webdataset
lpips
# mathutils==2.81.2
torch-fidelity
# lightning
huggingface_hub[cli,torch]
opencv-python
# xformers
